Rearrange navigator files
