+++ Sources/SwiftNavigationCoordinator/Navigator/Core/Modal/ModalState+Testing.swift
