+++ Sources/SwiftNavigationCoordinator/Navigator/Core/Modal/ModalState.swift
