+++ Sources/SwiftNavigationCoordinator/Navigator/Core/Modal/ModalStateDelegate.swift
