+++ Sources/SwiftNavigationCoordinator/Navigator/Core/Navigator+Modal.swift
