+++ Sources/SwiftNavigationCoordinator/Navigator/Core/Navigator+Specimen.swift
