+++ Sources/SwiftNavigationCoordinator/Navigator/Core/Navigator+Stack.swift
