+++ Sources/SwiftNavigationCoordinator/Navigator/Core/Specimen/SpecimenState+Testing.swift
