+++ Sources/SwiftNavigationCoordinator/Navigator/Core/Specimen/SpecimenState.swift
