+++ Sources/SwiftNavigationCoordinator/Navigator/Core/Stack/StackState+Testing.swift
