+++ Sources/SwiftNavigationCoordinator/Navigator/Core/Stack/StackState.swift
