+++ Sources/SwiftNavigationCoordinator/Navigator/Core/Stack/StackStateDelegate.swift
